[Test]:좌석 예약 동시성 test

+++ src/main/java/io/dev/concertreservationsystem/application/point_history/PointHistoryFacadeDTOResult.java
@@ -12,7 +12,7 @@ public record PointHistoryFacadeDTOResult(
         Long paymentId,
         String userId,
         PointTransactionType type,
-        Integer amount,
+        Long amount,
         Long resultPoint,
         LocalDateTime created_at,
         LocalDateTime updated_at
